+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 
 # 3.6.5
-- launch.json: Added "customVisualMemBlocks".
+- Debug command "-sjasmplus.path" added for dynamic loading of sld files during runtime (#156).
+- launch.json: Added "customVisualMemBlocks" (#157).
 
 # 3.6.4
 - launch.json: Added "load" support for .z80 files.